Luciferase-x5i.10: extract CollisionEngine + scoped Tetree collision-override refactor (RDR-008 P5)
Phase 5 of the RDR-008 AbstractSpatialIndex god-class decomposition. Extracts
the collision-detection cluster into a CollisionEngine feature object and
performs the §Decision-locked bounded Tetree collision-override refactor.

Architectural pattern (P3 refinement, LOCKED): per-cluster sub-interfaces in
the cluster's own package; the façade implements a private inner class that
delegates to the underlying methods. New seam:
  - collision.CollisionGeometry — 5 methods (findNodesIntersectingBounds +
    addNeighboringNodes subclass hooks, spatialRangeQuery composed query,
    two cached entity accessors). Naming-discipline exception for
    spatialRangeQuery is documented (the composed query is required because
    a direct call to the VolumeBounds subclass hook bypasses the
    doesNodeIntersectVolume filter — caught on first P5 verification pass).
  - collision.CollisionEngine — 7 public + ~22 private migrated verbatim;
    holds SpatialIndexCore + CollisionGeometry callback + Supplier<
    FineGrainedLockingStrategy> (late-bound for findCollisionsFineGrained,
    mirroring the KnnSearcher pattern).
  - Façade adds protected final CollisionEngine collisions field + eager ctor
    construction (this-escape invariant preserved — callback stores 'this'
    but the engine ctor never dispatches through it) + CollisionGeometryImpl
    private inner class with bidirectional NOTE blocks to FrustumGeometryImpl
    / CullGeometryImpl / KnnGeometryImpl documenting the intentional
    signature duplication (cluster-scoped sub-interfaces cannot share a base
    type without crossing package boundaries).
  - 7 facade public collision methods (checkCollision, findAllCollisions,
    findCollisions, findCollisionsFineGrained, findCollisionsInRegion,
    getCollisionShape, setCollisionShape) reduced to delegators.

Scoped Tetree refactor (BOUNDED — Phase 5 is the FIRST phase that breaks the
'subclasses byte-for-byte unchanged' invariant per RDR §Decision item 2):
Tetree's three collision overrides (findAllCollisions, findCollisions(ID),
findCollisionsInRegion) and the isEntityInAabt private helper now route
nucleus access through core.lock(), core.spatialIndex(), core.entityManager()
— the §Decision-sanctioned protected re-exposure path. Behavior identical;
only the access syntax changes. Octree, Prism, SFCArrayIndex are byte-for-
byte unchanged (git diff --stat -- octree/ prism/ sfc/ EMPTY).

Two bugs caught and fixed during stacked review (mem:feedback_review-stacking):
  - Tetree.findAllCollisions used HashSet<String> + createPairKey(hashCode
    ordering) for deduplication while the base sweep uses HashSet<
    UnorderedPair<ID>> (compareTo ordering). On hash collisions these
    disagree and the override would emit duplicate CollisionPair entries.
    Switched to UnorderedPair to align with the base contract; removed dead
    createPairKey helper.
  - TOCTOU window between super.findAllCollisions() (released lock) and the
    cross-tet sweep (re-acquired lock) — a concurrent writer could mutate
    the index in the gap. Fixed by acquiring the read lock BEFORE calling
    super.findAllCollisions() and holding it across both sections;
    ReentrantReadWriteLock allows the re-entrant read acquisition inside
    CollisionEngine.findAllCollisions.

Mandatory new test (RDR §Decision item 2 phase 5): added
TetreeCrossTetrahedraCollisionTest with two cases — single S0/S4 pair and
all 15 C(6,2) pairs across S0-S5 at level 21 (cellSize=1). The same-cell-
different-type collision path is unique to Tetree's tetrahedral subdivision
and invisible to the generic neighbor-traversal sweep that the base
findAllCollisions uses; without this test the override is unverified.

Verification:
  - Full lucien suite: 2455/0/0 green (post-clean).
  - Collision-cluster sweep (Octree+Tetree+Prism+cross-tet): 106/0/0.
  - First verification pass caught a Critical: CollisionEngine's initial
    findNodesIntersectingBounds(EntityBounds) wrapper routed directly to
    the subclass VolumeBounds hook, bypassing spatialRangeQuery's
    doesNodeIntersectVolume filter. OctreeCollisionAccuracyTest /
    OctreeCollisionDetectionTest / OctreeCollisionShapeIntegrationTest
    flagged the regression on the spanning-entity paths; fix re-routed
    through callback.spatialRangeQuery to match the pre-extraction
    facade wrapper.

Façade arc cumulative: 5851 → 3450 lines (−2401, 41% reduction; 5 of 6
phases done). G5 (Luciferase-x5i.11) gate is next; P6 (EntityLifecycleManager,
Luciferase-x5i.12) remains, which will additionally move the lockingStrategy
field into SpatialIndexCore and replace CollisionEngine's Supplier with
core.lockingStrategy() (P3 substantive-critic Significant#1 obligation).

Bead: Luciferase-x5i.10
RDR: RDR-008 (§Decision item 1 P3 refinement + item 2 phase 5)
